% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartseries.R
\name{chartseries}
\alias{chartseries}
\title{Interactive 'chart_Series'}
\usage{
chartseries(cs = NULL, width = NULL, height = NULL, elementId = NULL)
}
\description{
Convert \code{quantmod::chart_Series} to interactive d3fc htmlwidget chart.
}

