% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{html_dependency_d3fc}
\alias{html_dependency_d3fc}
\title{Dependencies for d3fc}
\usage{
html_dependency_d3fc(offline = TRUE)
}
\arguments{
\item{offline}{\code{logical} to use local file dependencies.  If \code{FALSE},
then the dependencies use the unpkg.com cdn as its \code{src}.}
}
\value{
\code{\link[htmltools]{htmlDependency}}
}
\description{
Dependencies for d3fc
}

